Aliased tmux and sudo, added compiler flags
See the comments in the file for explanations of the aliases. I turned
up the warnings in my flags because I pretty much always want to be
warned since I have the option of just ignoring them if, for example,
I'm compiling an unmodified third-party library that is known to work.